# services/idun_agent_web/schema/manager-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Idun Agent Manager API",
    "description": "Idun service for managing AI agents",
    "version": "0.5.3"
  },
  "paths": {
...
    "/api/v1/guardrails/": {
      "post": {
        "tags": [
          "Guardrails"
        ],
        "summary": "Create managed guardrail config",
        "description": "Create a new managed guardrail configuration.",
        "operationId": "create_guardrail_api_v1_guardrails__post",
        "parameters": [
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedGuardrailCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedGuardrailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Guardrails"
        ],
        "summary": "List managed guardrail configs",
        "description": "List managed guardrail configurations with pagination.",
        "operationId": "list_guardrails_api_v1_guardrails__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManagedGuardrailRead"
                  },
                  "title": "Response List Guardrails Api V1 Guardrails  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guardrails/{id}": {
      "get": {
        "tags": [
          "Guardrails"
        ],
        "summary": "Get managed guardrail config by ID",
        "description": "Get a managed guardrail configuration by ID.",
        "operationId": "get_guardrail_api_v1_guardrails__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedGuardrailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Guardrails"
        ],
        "summary": "Delete managed guardrail config",
        "description": "Delete a managed guardrail configuration permanently.",
        "operationId": "delete_guardrail_api_v1_guardrails__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Guardrails"
        ],
        "summary": "Update managed guardrail config",
        "description": "Update a guardrail configuration.",
        "operationId": "patch_guardrail_api_v1_guardrails__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-project-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "x-workspace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedGuardrailPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedGuardrailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManagedGuardrailCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "guardrail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimpleBanListConfig"
              },
              {
                "$ref": "#/components/schemas/SimplePIIConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleNSFWTextConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleToxicLanguageConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleGibberishTextConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleBiasCheckConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleCompetitionCheckConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleCorrectLanguageConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleRestrictToTopicConfig"
              }
            ],
            "title": "Guardrail",
            "description": "Guardrail configuration"
          }
        },
        "type": "object",
        "required": [
          "name",
          "guardrail"
        ],
        "title": "ManagedGuardrailCreate",
        "description": "Create managed guardrail model for requests."
      },
      "ManagedGuardrailPatch": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "guardrail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimpleBanListConfig"
              },
              {
                "$ref": "#/components/schemas/SimplePIIConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleNSFWTextConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleToxicLanguageConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleGibberishTextConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleBiasCheckConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleCompetitionCheckConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleCorrectLanguageConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleRestrictToTopicConfig"
              }
            ],
            "title": "Guardrail",
            "description": "Guardrail configuration"
          }
        },
        "type": "object",
        "required": [
          "name",
          "guardrail"
        ],
        "title": "ManagedGuardrailPatch",
        "description": "Full replacement schema for PUT of a managed guardrail."
      },
      "ManagedGuardrailRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "guardrail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimpleBanListConfig"
              },
              {
                "$ref": "#/components/schemas/SimplePIIConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleNSFWTextConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleToxicLanguageConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleGibberishTextConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleBiasCheckConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleCompetitionCheckConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleCorrectLanguageConfig"
              },
              {
                "$ref": "#/components/schemas/SimpleRestrictToTopicConfig"
              }
            ],
            "title": "Guardrail",
            "description": "Guardrail configuration"
          },
          "agent_count": {
            "type": "integer",
            "title": "Agent Count",
            "description": "Number of agents using this guardrail",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "guardrail",
          "created_at",
          "updated_at"
        ],
        "title": "ManagedGuardrailRead",
        "description": "Complete managed guardrail model for responses."
      },
...
      "SimpleBanListConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "ban_list",
            "title": "Config Id",
            "default": "ban_list"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "banned_words": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Banned Words",
            "description": "A list of strings (words or phrases) to block"
          }
        },
        "type": "object",
        "required": [
          "banned_words"
        ],
        "title": "SimpleBanListConfig"
      },
      "SimpleBiasCheckConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "bias_check",
            "title": "Config Id",
            "default": "bias_check"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "description": "Sensitivity level for bias detection"
          }
        },
        "type": "object",
        "required": [
          "threshold"
        ],
        "title": "SimpleBiasCheckConfig"
      },
      "SimpleCompetitionCheckConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "competition_check",
            "title": "Config Id",
            "default": "competition_check"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "competitors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Competitors",
            "description": "Names of competitor companies or products"
          }
        },
        "type": "object",
        "required": [
          "competitors"
        ],
        "title": "SimpleCompetitionCheckConfig"
      },
      "SimpleCorrectLanguageConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "correct_language",
            "title": "Config Id",
            "default": "correct_language"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "expected_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expected Languages",
            "description": "Valid ISO language codes (e.g., en, fr, es)"
          }
        },
        "type": "object",
        "required": [
          "expected_languages"
        ],
        "title": "SimpleCorrectLanguageConfig"
      },
      "SimpleGibberishTextConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "gibberish_text",
            "title": "Config Id",
            "default": "gibberish_text"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "description": "Sensitivity level for gibberish detection"
          }
        },
        "type": "object",
        "required": [
          "threshold"
        ],
        "title": "SimpleGibberishTextConfig"
      },
      "SimpleNSFWTextConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "nsfw_text",
            "title": "Config Id",
            "default": "nsfw_text"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "description": "Sensitivity level for NSFW content"
          }
        },
        "type": "object",
        "required": [
          "threshold"
        ],
        "title": "SimpleNSFWTextConfig"
      },
      "SimplePIIConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "detect_pii",
            "title": "Config Id",
            "default": "detect_pii"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "pii_entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pii Entities",
            "description": "List of PII entity types to detect"
          }
        },
        "type": "object",
        "required": [
          "pii_entities"
        ],
        "title": "SimplePIIConfig"
      },
      "SimpleRestrictToTopicConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "restrict_to_topic",
            "title": "Config Id",
            "default": "restrict_to_topic"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics",
            "description": "List of allowed topics"
          }
        },
        "type": "object",
        "required": [
          "topics"
        ],
        "title": "SimpleRestrictToTopicConfig"
      },
      "SimpleToxicLanguageConfig": {
        "properties": {
          "config_id": {
            "type": "string",
            "const": "toxic_language",
            "title": "Config Id",
            "default": "toxic_language"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          },
          "reject_message": {
            "type": "string",
            "title": "Reject Message",
            "default": ""
          },
          "threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "description": "Sensitivity level for toxic language"
          }
        },
        "type": "object",
        "required": [
          "threshold"
        ],
        "title": "SimpleToxicLanguageConfig"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}